Break up admin tool into modules.
